% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert between Ether units}
\usage{
convert(x, from = "wei", to)
}
\arguments{
\item{x}{Numeric. A vector giving values to be converted.}

\item{from}{Character. The units of Ether in which \code{x} is given. See
\code{\link{eth_units}} for valid values. Default is \code{'wei'}.}

\item{to}{Character. The target unit of Ether to which \code{x} will be
converted. See \code{\link{eth_units}} for valid values.}
}
\value{
A vector giving \code{x} in units specified by \code{to}.
}
\description{
Convert between Ether units.
}
\seealso{
\code{\link{eth_units}}
}
\keyword{ETH}
\keyword{Ethereum,}
\keyword{currency,}
\keyword{units,}
