% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abi.R
\name{get_abi}
\alias{get_abi}
\title{Return the ABI of a contract}
\usage{
get_abi(address, api_key)
}
\arguments{
\item{address}{Character. A single ethereum contract address as a character
string (40 hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key (see Details).}
}
\value{
An \code{abi} object (inherits \code{json}) giving the ABI of the
contract.
}
\description{
Check whether a contract has been verified by Etherscan.
}
\details{
\code{get_abi} uses the Etherscan API. Register for an API key
at the \href{https://etherscan.io/apis}{Etherscan Developer APIs page}.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer
APIs are provided as a community service and without warranty, so please just
use what you need and no more. They support both GET/POST requests and a rate
limit of 5 requests/sec.}
}

\seealso{
\code{\link[=parse_abi]{parse_abi()}}
}
\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{contract,}
\keyword{crypto,}
\keyword{cryptocurrency,}
