% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balance.R
\name{get_balance}
\alias{get_balance}
\title{Get Ethereum balance for an address}
\usage{
get_balance(address, api_key)
}
\arguments{
\item{address}{Character. A character vector of one or more ethereum addresses (40
hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key (see Details).}
}
\value{
A numeric vector of balances (in ether units) corresponding to the
  elements of \code{address}.
}
\description{
Get normal or internal transactions for an Ethereum address.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs are
provided as a community service and without warranty, so please just use what
you need and no more. They support both GET/POST requests and a rate limit of
5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{crypto,}
\keyword{cryptocurrency,}
\keyword{transaction,}
