% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_events.R
\name{get_events}
\alias{get_events}
\title{Get event logs}
\usage{
get_events(address, topic0, fromBlock, toBlock, api_key)
}
\arguments{
\item{address}{Character. A single Ethereum contract address as a character
string (40 hexadecimal characters prepended by '0x').}

\item{topic0}{Character. The topic0 keccak256 hash.}

\item{fromBlock}{Numeric. The minimum block number of interest.}

\item{toBlock}{Numeric. The maximum block number of interest. Also accepts
\code{'latest'}.}

\item{api_key}{An Etherscan API key (see Details).}
}
\description{
Get event logs for a given contract address and topic
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs
are provided as a community service and without warranty, so please just use
what you need and no more. They support both GET/POST requests and a rate
limit of 5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{crypto,}
\keyword{cryptocurrency,}
\keyword{transaction,}
