% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tx.R
\name{get_tx}
\alias{get_tx}
\title{get_tx}
\usage{
get_tx(tx, api_key)
}
\arguments{
\item{tx}{A character vector of one or more transaction hashes (hexadecimal,
64 characters, with 0x prepended).}

\item{api_key}{An Etherscan API key (see Details).}
}
\value{
A \code{data.frame} with one row per element of \code{tx}, with
columns: \code{blockHash}, \code{blockNumber}, \code{from}, \code{gas},
\code{gasPrice}, \code{hash}, \code{input}, \code{nonce}, \code{to},
\code{transactionIndex}, \code{value}, \code{v}, \code{r}, \code{s}.
}
\description{
Get details of one or more transactions by hash.
}
\details{
\code{get_tx} uses the Etherscan API to source information about
transactions to and from an Ethereum address. Register for an API key at
the \href{https://etherscan.io/apis}{Etherscan Developer APIs page}.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs are
provided as a community service and without warranty, so please just use what
you need and no more. They support both GET/POST requests and a rate limit of
5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{crypto,}
\keyword{cryptocurrency,}
\keyword{transaction,}
