% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txs.R
\name{get_txs}
\alias{get_txs}
\title{Get transactions for an address}
\usage{
get_txs(address, api_key, internal = FALSE, no_errors = TRUE)
}
\arguments{
\item{address}{Character. A single ethereum address as a character string (40
hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key (see Details).}

\item{internal}{Logical. Should normal (\code{FALSE}, default) or internal
(\code{TRUE}) transactions be queried?}

\item{no_errors}{Logical. Should unsuccessful transactions be omitted
(\code{FALSE}, default)?}
}
\value{
If \code{internal} is \code{FALSE}, a \code{tbl_df} with the
  following elements:
 \itemize{
   \item{\code{blockNumber (<dbl>)}}{}
   \item{\code{timeStamp (<dttm>)}}{}
   \item{\code{hash (<chr>)}}{}
   \item{\code{nonce (<dbl>)}}{}
   \item{\code{blockHash (<chr>)}}{}
   \item{\code{transactionIndex (<dbl>)}}{}
   \item{\code{from (<chr>)}}{}
   \item{\code{to (<chr>)}}{}
   \item{\code{value (<dbl>)}}{}
   \item{\code{gas (<dbl>)}}{}
   \item{\code{gasPrice (<dbl>)}}{}
   \item{\code{isError (<dbl>)}}{}
   \item{\code{txreceipt_status (<chr>)}}{}
   \item{\code{input (<chr>)}}{}
   \item{\code{contractAddress (<chr>)}}{}
 }
 If \code{internal} is \code{TRUE}, a \code{tbl_df} with the following
 elements:
 \itemize{
   \item{\code{blockNumber (<dbl>)}}{}
   \item{\code{timeStamp (<dttm>)}}{}
   \item{\code{hash (<chr>)}}{}
   \item{\code{from (<chr>)}}{}
   \item{\code{to (<chr>)}}{}
   \item{\code{value (<dbl>)}}{}
   \item{\code{contractAddress (<chr>)}}{}
   \item{\code{input (<chr>)}}{}
   \item{\code{type (<chr>)}}{}
   \item{\code{gas (<dbl>)}}{}
   \item{\code{gasUsed (<dbl>)}}{}
   \item{\code{traceId (<dbl>)}}{}
   \item{\code{isError (<dbl>)}}{}
   \item{\code{errCode (<chr>)}}{}
   \item{\code{value_eth (<dbl>)}}{}
}
}
\description{
Get normal or internal transactions for an Ethereum address.
}
\details{
\code{get_txs} uses the Etherscan API to source information about
  transactions to and from an Ethereum address. Register for an API key at
  the \href{https://etherscan.io/apis}{\emph{Etherscan Developer APIs page}}.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs are
provided as a community service and without warranty, so please just use what
you need and no more. They support both GET/POST requests and a rate limit of
5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{crypto,}
\keyword{cryptocurrency,}
\keyword{transaction,}
