% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Eth_Addr_Inof.R
\name{getBalance}
\alias{getBalance}
\title{getBalance Get Eth addr balance}
\usage{
getBalance(eth_addr)
}
\arguments{
\item{eth_addr}{}
}
\value{
Eth addr balance
}
\description{
getBalance Get Eth addr balance
}
