% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Eth_Addr_Inof.R
\name{getInfo}
\alias{getInfo}
\title{getInfo Get the balance and the real value of a ETH address}
\usage{
getInfo(eth_addr, name = "unamed", use_cny = T)
}
\arguments{
\item{eth_addr}{the address}

\item{name}{name of the address}

\item{use_cny}{True for cny, False for usd}
}
\value{
a one line data.frame
}
\description{
getInfo Get the balance and the real value of a ETH address
}
