% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Eth_Addr_Inof.R
\name{getTrans}
\alias{getTrans}
\title{getTrans Get the transaction histroy of an ETH address}
\usage{
getTrans(eth_addr, number = 10)
}
\arguments{
\item{eth_addr}{the address}

\item{number}{number of lines}
}
\value{
the data.frame of all transactions
}
\description{
getTrans Get the transaction histroy of an ETH address
}
