% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_out.R
\name{in_out}
\alias{in_out}
\title{Get the name of a verified contract}
\usage{
in_out(address, api_key, get_names = FALSE)
}
\arguments{
\item{address}{Character. A single verified ethereum contract address as a
character string (40 hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key.}

\item{get_names}{Logical. Get contract names for verified contracts?}
}
\value{
A summary of inflow and outflow of ETH from an address.
}
\description{
Return the name of a verified contract.
}
\seealso{
\code{\link[=get_abi]{get_abi()}}
}
\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{contract,}
\keyword{crypto,}
\keyword{cryptocurrency,}
