% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_contract.R
\name{is_contract}
\alias{is_contract}
\title{Check whether an address is a contract address}
\usage{
is_contract(address, api_key)
}
\arguments{
\item{address}{Character. A single ethereum address as a character string (40
hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key (see Details).}
}
\value{
Logical. \code{TRUE} if the address is a contract address.
}
\description{
Check whether an Ethereum address is a contract address or an external
address.
}
\details{
\code{is_contract} uses the Etherscan API. Register for an API key
  at the \href{https://etherscan.io/apis}{\emph{Etherscan Developer APIs
  page}}.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs are
provided as a community service and without warranty, so please just use what
you need and no more. They support both GET/POST requests and a rate limit of
5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{contract,}
\keyword{crypto,}
\keyword{cryptocurrency,}
