% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_verified.R
\name{is_verified}
\alias{is_verified}
\title{Check whether a contract has verified source code}
\usage{
is_verified(address, api_key)
}
\arguments{
\item{address}{Character. A single ethereum contract address as a character
string (40 hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key (see Details).}
}
\value{
Logical. \code{TRUE} if the contract is verified.
}
\description{
Check whether a contract has been verified by Etherscan.
}
\details{
\code{is_verified} uses the Etherscan API. Register for an API key
at the \href{https://etherscan.io/apis}{Etherscan Developer APIs page}.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs
are provided as a community service and without warranty, so please just use
what you need and no more. They support both GET/POST requests and a rate
limit of 5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{contract,}
\keyword{crypto,}
\keyword{cryptocurrency,}
