% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_txs.R
\name{get_txs}
\alias{get_txs}
\title{Get transactions for an address}
\usage{
get_txs(
  address,
  api_key,
  internal = FALSE,
  type = c("normal", "internal", "ERC20", "ERC721"),
  startblock = 0,
  endblock = 999999999,
  network = "mainnet",
  first_page = FALSE,
  no_errors = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{address}{Character. A single ethereum address as a character string (40
hexadecimal characters prepended by '0x').}

\item{api_key}{An Etherscan API key (see Details).}

\item{internal}{Logical. Should normal (\code{FALSE}, default) or internal
(\code{TRUE}) transactions be queried? \emph{Deprecated. Please use \code{type} instead.}}

\item{type}{The type of transaction to query. One of \code{'normal'},
\code{'internal'}, \code{'ERC20'}, or \code{'ERC721'}.}

\item{startblock}{Starting block for transaction query. Default is 0.}

\item{endblock}{Ending block for transaction query. Default is 999999999.}

\item{network}{Ethereum network to use. One of \code{'mainnet'} (default),
\code{'ropsten'}, \code{'rinkeby'}, \code{'kovan'}, or \code{'goerli'}.}

\item{first_page}{Logical. Should only the first page of results (up to
10,000 transactions) be returned.}

\item{no_errors}{Logical. Should unsuccessful transactions be omitted
(\code{FALSE}, default)?}

\item{quiet}{Logical. Suppress messages? Default is \code{FALSE}.}
}
\value{
A \code{tbl_df} with transaction details. Columns vary depending on
\code{type}.
}
\description{
Get normal, internal, ERC-20, or ERC-721 transactions for an Ethereum address.
}
\details{
\code{get_txs} uses the Etherscan API to source information about
transactions to and from an Ethereum address. Register for an API key at
the \href{https://etherscan.io/apis}{Etherscan Developer APIs page}.
}
\section{Warning}{

As per the Etherscan documentation, \emph{the Etherscan Ethereum Developer APIs
are provided as a community service and without warranty, so please just use
what you need and no more. They support both GET/POST requests and a rate
limit of 5 requests/sec.}
}

\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{crypto,}
\keyword{cryptocurrency,}
\keyword{transaction,}
