% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_out.R
\name{in_out}
\alias{in_out}
\title{Summarize ETH flow for an address}
\usage{
in_out(address, api_key, get_names = FALSE)
}
\arguments{
\item{address}{Character. A single externally owned account (user Ethereum
address).}

\item{api_key}{An Etherscan API key (see Details).}

\item{get_names}{Logical. Get contract names for verified contracts?}
}
\value{
A summary of inflow and outflow of ETH from an address.
}
\description{
Summarize inflow and outflow of ETH from an address.
}
\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{contract,}
\keyword{crypto,}
\keyword{cryptocurrency,}
