% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_abi.R
\name{parse_abi}
\alias{parse_abi}
\title{Parse an ABI}
\usage{
parse_abi(abi)
}
\arguments{
\item{abi}{An \code{abi} object returned by \code{\link{get_abi}}.}
}
\value{
A \code{tbl_df} with the following columns:
  \itemize{
    \item{\code{constant (<lgl>)}}{}
    \item{\code{name (<chr>)}}{}
    \item{\code{input_indexed (<chr>)}}{}
    \item{\code{input_name (<chr>)}}{}
    \item{\code{input_type (<chr>)}}{}
    \item{\code{output_name (<chr>)}}{}
    \item{\code{output_type (<chr>)}}{}
    \item{\code{payable (<lgl>)}}{}
    \item{\code{stateMutability (<chr>)}}{}
    \item{\code{type (<chr>)}}{}
    \item{\code{anonymous (<lgl>)}}{}
    \item{\code{signature (<chr>)}}{}
    \item{\code{method (<chr>)}}{}
  }
}
\description{
Parse an \code{abi} object returned by \code{\link{get_abi}}.
}
\details{
The values given in the \code{method} column are the first 8
  characters (4 bytes) of the keccak256 hashes of the method signatures.
}
\seealso{
\code{\link{get_abi}} \code{\link{keccak256}}
}
\keyword{ETH}
\keyword{Ethereum,}
\keyword{blockchain,}
\keyword{contract,}
\keyword{crypto,}
\keyword{cryptocurrency,}
